requirements.txt
PyQt6>=6.6.0
PyQt6-WebEngine>=6.6.0
requests>=2.31.0
Pillow>=10.0.0